Removed many includes of array.hpp
