+++ include/dynd/array.hpp
@@ -1110,4 +1110,32 @@ namespace nd {
   DYND_API array combine_into_tuple(size_t field_count, const array *field_values);
 
 } // namespace dynd::nd
+
+namespace ndt {
+
+  /**
+   * Does a value lookup into an array of type "N * T", without
+   * bounds checking the index ``i`` or validating that ``a`` has the
+   * required type. Use only when these checks have been done externally.
+   */
+  template <typename T>
+  inline const T &unchecked_fixed_dim_get(const nd::array &a, intptr_t i)
+  {
+    const size_stride_t *md = reinterpret_cast<const size_stride_t *>(a.get()->metadata());
+    return *reinterpret_cast<const T *>(a.cdata() + i * md->stride);
+  }
+
+  /**
+   * Does a writable value lookup into an array of type "N * T", without
+   * bounds checking the index ``i`` or validating that ``a`` has the
+   * required type. Use only when these checks have been done externally.
+   */
+  template <typename T>
+  inline T &unchecked_fixed_dim_get_rw(const nd::array &a, intptr_t i)
+  {
+    const size_stride_t *md = reinterpret_cast<const size_stride_t *>(a.get()->metadata());
+    return *reinterpret_cast<T *>(a.data() + i * md->stride);
+  }
+
+} // namespace dynd::ndt
 } // namespace dynd
