+++ include/dynd/types/callable_type.hpp
@@ -5,7 +5,6 @@
 
 #pragma once
 
-#include <dynd/array.hpp>
 #include <dynd/types/fixed_dim_type.hpp>
 #include <dynd/types/tuple_type.hpp>
 #include <dynd/types/struct_type.hpp>
