+++ include/dynd/types/dim_fragment_type.hpp
@@ -8,7 +8,7 @@
 #include <vector>
 #include <string>
 
-#include <dynd/array.hpp>
+#include <dynd/shortvector.hpp>
 #include <dynd/types/base_dim_type.hpp>
 
 namespace dynd {
