+++ include/dynd/types/ellipsis_dim_type.hpp
@@ -8,7 +8,6 @@
 #include <vector>
 #include <string>
 
-#include <dynd/array.hpp>
 #include <dynd/types/base_dim_type.hpp>
 
 namespace dynd {
